refactor(services): migrate flow domain services to flow/ subpackage
Phase 3-3: Create flow/ subpackage following task/, pr/, issue/ pattern.

Changes:
- Create src/vibe3/services/flow/ subpackage with 18 modules
- Migrate flow_*.py files to flow/*.py using git mv
- Update all internal imports to use new paths (flow.service, flow.cleanup, etc.)
- Create 12 re-export shims for backward compatibility
- Update services/__init__.py TYPE_CHECKING and _SYMBOL_MODULES
- Update test fixtures and patches to use new import paths

Files migrated (18):
- flow_service.py → flow/service.py
- flow_cleanup_service.py → flow/cleanup.py
- flow_recovery_service.py → flow/recovery.py
- flow_status_service.py → flow/status.py
- flow_timeline_service.py → flow/timeline.py
- flow_projection_service.py → flow/projection.py
- flow_status_resolver.py → flow/status_resolver.py
- flow_resume_resolver.py → flow/resume_resolver.py
- flow_reader.py → flow/reader.py
- flow_classifier.py → flow/classifier.py
- flow_consistency_check.py → flow/consistency.py
- flow_rebuild_usecase.py → flow/rebuild.py
- flow_rebuild_postconditions.py → flow/rebuild_postconditions.py
- flow_factory.py → flow/factory.py
- flow_block_mixin.py → flow/block_mixin.py
- flow_read_mixin.py → flow/read_mixin.py
- flow_write_mixin.py → flow/write_mixin.py
- flow_transition.py → flow/transition.py

Files retained at top level (2):
- flow_orchestrator_service.py (cross-domain orchestrator)
- abandon_flow_service.py (cross-domain orchestrator)

Re-export shims created (12):
- All 9 originally planned shims
- Plus 3 additional shims for test compatibility:
  * flow_consistency_check.py
  * flow_rebuild_usecase.py
  * flow_rebuild_postconditions.py

All backward-compatible imports work:
- from vibe3.services import FlowService
- from vibe3.services.flow import FlowService
- from vibe3.services.flow.service import FlowService
- from vibe3.services.flow_service import FlowService (shim)

Verification:
- All 258 targeted tests pass
- mypy: no errors
- ruff: all checks pass
- No circular dependencies

Issue: #2107
