+++ src/vibe3/services/flow/block_mixin.py
