+++ src/vibe3/services/flow/reader.py
@@ -1,0 +1,5 @@
+"""Re-export shim — FlowReader has moved to clients.protocols.flow."""
+
+from vibe3.clients import FlowReader
+
+__all__ = ["FlowReader"]
