+++ src/vibe3/services/flow/resume_resolver.py
@@ -1,0 +1,67 @@
+"""Flow resume resolver service.
+
+Provides pure logic to infer the correct GitHub state label for a flow based
+on its local reference states (pr_ref, audit_ref, plan_ref, report_ref).
+"""
+
+from vibe3.models import FlowState, IssueState, VerdictValue
+from vibe3.services.verdict_policy import passes_review
+
+
+def infer_resume_label(flow_state: FlowState) -> IssueState:
+    """Infer the correct state label a flow should return to.
+
+    Used by:
+    1. Qualify gate automatic unblock (determining target label)
+    2. `vibe3 task resume --label` (without explicit value)
+
+    Priority Rules (High to Low):
+    1. pr_ref exists -> HANDOFF (Code is ready for delivery)
+    2. latest_verdict exists -> MERGE_READY / IN_PROGRESS / HANDOFF (Based on verdict)
+    3. report_ref exists -> REVIEW (Code is written, needs review)
+    4. plan_ref exists -> IN_PROGRESS (Plan exists, needs execution)
+    5. default -> READY (Ready to be picked up)
+
+    Args:
+        flow_state: The current local state of the flow
+
+    Returns:
+        The inferred target IssueState
+    """
+    if flow_state.pr_ref:
+        # PR already exists -> Manager should take over
+        return IssueState.HANDOFF
+
+    if flow_state.latest_verdict:
+        verdict: VerdictValue = flow_state.latest_verdict.verdict
+        # Verdict is the primary signal for review completion.
+        if passes_review(verdict):
+            # PASS -> merge-ready immediately
+            if verdict == "PASS":
+                return IssueState.MERGE_READY
+            # MINOR requires audit_ref
+            if verdict == "MINOR":
+                if flow_state.audit_ref:
+                    return IssueState.MERGE_READY
+                # Missing audit_ref -> back to review
+                return IssueState.REVIEW
+        else:
+            # blocks_merge(verdict) is True for these
+            if verdict in {"MAJOR", "BLOCK"}:
+                if flow_state.audit_ref:
+                    return IssueState.IN_PROGRESS
+                # Missing audit_ref -> back to review
+                return IssueState.REVIEW
+            # REFUSE / UNKNOWN -> manager takes over
+            return IssueState.HANDOFF
+
+    if flow_state.plan_ref and flow_state.report_ref:
+        # Code has been written (report exists) but no review yet
+        return IssueState.REVIEW
+
+    if flow_state.plan_ref:
+        # Plan exists but no code (report) yet
+        return IssueState.IN_PROGRESS
+
+    # No plan_ref -> ready to be picked up by executor
+    return IssueState.READY
