+++ src/vibe3/services/flow_consistency_check.py
@@ -1,111 +1,15 @@
-"""Shared flow/worktree/ref consistency checks."""
+"""Re-export shim — flow consistency moved to vibe3.services.flow.consistency."""
 
-from __future__ import annotations
+from vibe3.services.flow.consistency import (
+    FlowConsistencyCode,
+    FlowConsistencyResult,
+    apply_consistency_fix,
+    check_flow_consistency,
+)
 
-from dataclasses import dataclass
-from enum import StrEnum
-from typing import Any, Mapping
-
-from vibe3.services.shared.paths import check_ref_exists
-
-
-class FlowConsistencyCode(StrEnum):
-    """Machine-readable consistency failure codes."""
-
-    OK = "ok"
-    MISSING_WORKTREE = "missing_worktree"
-    MISSING_RECORDED_WORKTREE = "missing_recorded_worktree"
-    MISSING_REF = "missing_ref"
-
-
-@dataclass(frozen=True)
-class FlowConsistencyResult:
-    """Result of shared structural consistency detection.
-
-    Classification:
-    - needs_rebuild: physical scene must be destroyed and recreated
-    - fix_action: cheap local fix (e.g. backfill DB field), no rebuild needed
-    - both False: scene is consistent
-    """
-
-    needs_rebuild: bool
-    code: FlowConsistencyCode = FlowConsistencyCode.OK
-    reason: str = ""
-    severity: str = ""
-    ref_field: str | None = None
-    ref_value: str | None = None
-    fix_action: str | None = None
-    worktree_path: str | None = None
-
-
-def check_flow_consistency(
-    branch: str,
-    flow_state: Mapping[str, Any],
-    *,
-    git_client: Any,
-) -> FlowConsistencyResult:
-    """Check whether a flow scene is structurally consistent.
-
-    Classification:
-    - MISSING_WORKTREE: physical worktree gone -> needs full rebuild
-    - MISSING_RECORDED_WORKTREE: worktree exists but DB out of sync -> cheap fix
-    - MISSING_REF: artifact ref points to non-existent file -> needs rebuild
-    - OK: scene is consistent
-    """
-    worktree_path = git_client.find_worktree_path_for_branch(branch)
-    if worktree_path is None:
-        return FlowConsistencyResult(
-            needs_rebuild=True,
-            code=FlowConsistencyCode.MISSING_WORKTREE,
-            reason=f"Worktree does not exist for branch '{branch}'",
-            severity="critical",
-        )
-
-    if branch.startswith("task/issue-") and not flow_state.get("worktree_path"):
-        # Worktree physically exists but DB doesn't know about it.
-        # This is a cheap fix: just backfill the DB field.
-        return FlowConsistencyResult(
-            needs_rebuild=False,
-            code=FlowConsistencyCode.MISSING_RECORDED_WORKTREE,
-            reason="Worktree exists but is not recorded in flow_state",
-            severity="low",
-            fix_action="backfill_worktree_path",
-            worktree_path=str(worktree_path),
-        )
-
-    for ref_field in ("plan_ref", "report_ref", "audit_ref"):
-        ref_value = flow_state.get(ref_field)
-        if not ref_value:
-            continue
-        _, exists = check_ref_exists(
-            str(ref_value),
-            branch,
-            git_client=git_client,
-        )
-        if not exists:
-            return FlowConsistencyResult(
-                needs_rebuild=True,
-                code=FlowConsistencyCode.MISSING_REF,
-                reason=f"{ref_field} file not found: {ref_value}",
-                severity="critical",
-                ref_field=ref_field,
-                ref_value=str(ref_value),
-            )
-
-    return FlowConsistencyResult(needs_rebuild=False)
-
-
-def apply_consistency_fix(
-    result: FlowConsistencyResult,
-    branch: str,
-    *,
-    store: Any,
-) -> bool:
-    """Apply a cheap fix identified by check_flow_consistency.
-
-    Returns True if fix was applied, False if no fix applicable.
-    """
-    if result.fix_action == "backfill_worktree_path" and result.worktree_path:
-        store.update_flow_state(branch, worktree_path=result.worktree_path)
-        return True
-    return False
+__all__ = [
+    "FlowConsistencyCode",
+    "FlowConsistencyResult",
+    "apply_consistency_fix",
+    "check_flow_consistency",
+]
