+++ tests/vibe3/services/test_check_merge_fix.py
@@ -214,7 +214,7 @@ class TestStaleFlowHandling:
             "task_issue_number": None,
         }
         mock_store.get_issue_links.return_value = [
-            {"issue_number": 431, "issue_role": "task"}
+            {"branch": "task/issue-431", "issue_number": 431, "issue_role": "task"}
         ]
         mock_github_client.view_issue.return_value = {
             "number": 431,
@@ -224,9 +224,7 @@ class TestStaleFlowHandling:
         }
         mock_github_client.list_prs_for_branch.return_value = []
 
-        with patch(
-            "vibe3.services.flow_rebuild_usecase.FlowRebuildUsecase"
-        ) as rebuild_cls:
+        with patch("vibe3.services.flow.rebuild.FlowRebuildUsecase") as rebuild_cls:
             rebuild = rebuild_cls.return_value
             rebuild.rebuild_issue_flow.return_value = {
                 "branch": "task/issue-431",
